NFC: Move usdt helper routines out of utils
This commit just cut and pastes code around. It's nicer to have
small, orthogonal modules.
