+++ src/ast/semantic_analyser.cpp
@@ -6,7 +6,7 @@
 #include "parser.tab.hh"
 #include "printf.h"
 #include "tracepoint_format_parser.h"
-#include "utils.h"
+#include "usdt.h"
 #include <algorithm>
 #include <cstring>
 #include <regex>
