+++ src/utils.h
@@ -18,16 +18,6 @@ struct vmlinux_location
   bool raw;         // file is either as ELF (false) or raw BTF data (true)
 };
 extern const struct vmlinux_location vmlinux_locs[];
-
-struct usdt_probe_entry
-{
-  std::string path;
-  std::string provider;
-  std::string name;
-  int num_locations;
-};
-typedef std::vector<usdt_probe_entry> usdt_probe_list;
-
 class MountNSException : public std::exception
 {
 public:
@@ -91,20 +81,6 @@ public:
   {
     ofile = stdout;
   }
-};
-
-class USDTHelper
-{
-public:
-  static usdt_probe_entry find(int pid,
-                               const std::string &target,
-                               const std::string &provider,
-                               const std::string &name);
-  static usdt_probe_list probes_for_provider(const std::string &provider);
-  static usdt_probe_list probes_for_pid(int pid);
-  static usdt_probe_list probes_for_path(const std::string &path);
-  static void read_probes_for_pid(int pid);
-  static void read_probes_for_path(const std::string &path);
 };
 
 // Hack used to suppress build warning related to #474
